+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Added API Backend in wide and ultra view console get view
+- Added the option to filter for specific backendBasepath
+
 
 ## [1.0.1] 2020-06-17
 ### Fixed